llm-selenium-agent/package/requirements.txt
streamlit
tqdm
selenium
pandas
numpy
webdriver_manager
twocaptcha
python-dotenv
pillow
beautifulsoup4
pyyaml

